Type-disagreement repair 2: 833 -> 677, merged tree 695 -> 586 TUs, flash payload BYTE-IDENTICAL
Prioritised by translation-unit yield, arbitrated by the shipped image, and
constrained to size-neutral repairs because Renode belonged to another agent.

YIELD ORDERING, COMPUTED NOT INHERITED
Re-ran Stage 04's own partition (t04_cohesive_tu imported READ-ONLY) over the
1,615 retained app sources: baseline 695 TUs, ceiling with the `type` rule off
243 (Stage 04's "120" is stale; the brief's 452 distance is right). Greedy
per-symbol yield: get_device_info 27, assert_post_action 23, memset_bytes 16,
memcpy 5 -- the four named symbols confirmed as the top, worth 71 of the 452.
299 symbols are the sole blocker of at least one adjacency.

THREE NEW CALLEE-ORACLE CALIBRATION BUGS, each found by disassembling a
top-yield symbol before trusting the tool:
- in_stack still fires on a function's OWN LOCALS when frame == 0
  (__aeabi_uldivmod -> "arity 8"). Rule adopted: trust in_stack only when
  frame > 0 and slots start at 0.
- `eors r0,r0` counts as an argument read -- this made assert_post_action
  (23 TUs) look like arity >= 1. 3 functions affected.
- Tail-call thunks hide pass-through arguments (movs r3,#0; b.w ...): 94 in the
  image; the three log thunks are arity 3, not 0.
- Also: ret_r1_expl alone is NOT evidence of a 64-bit return (z_spin_unlock_valid
  writes r1 as scratch on every path); an earlier canonicaliser emitted
  `uint64_t z_spin_unlock_valid(uint32_t)` off it.

GATE USED, since Renode was unavailable: PER-OBJECT BYTE IDENTITY. Anything whose
object changed was REVERTED, not rationalised.

RESULTS (authoritative from-scratch build)
nm -u 0; dup globals 0; pin gates app 627 bound / 0 escaping / 0 unknown, net 0/0
on the frozen ELF; check_thread_create_stack_args 10/10; verify_data 995/995.
FLASH 956,480 B / 97.35 % -- DELTA 0; RAM delta 0; every sub-image delta 0;
.text bytes cmp-identical.
Per-object: 2,837 matched, 2,836 identical, 1 differs -- uname.c.obj, 4 bytes of
embedded build clock, gc-sectioned out.
app_update.bin: 957,072 of 957,328 bytes identical; the 256-byte tail is the
randomised RSA-PSS signature. THE FLASH PAYLOAD IS BYTE-IDENTICAL TO HEAD'S.
Stage 03 agreement test: 833 -> 677 disagreeing (156 fixed; the previous pass
managed 104 of 1,014).
Stage 04 partition: 695 -> 586 TUs -- 109 TUs, 24 % of the distance, for ZERO BYTES.
Four-way split: FIXED 156 / CONFIRMED REAL 144 / FALSE ALARM 168 / UNSETTLED 365.

CORRECTIONS AGAINST ITS OWN INTEREST
- Rounds 1-3 gated against an INCREMENTAL build dir where `cp -a` restores gave
  ninja stale objects; one object differed for three rounds from a discarded
  plan. All headline numbers were re-derived from a from-scratch build.
- The previous pass's "6 short definitions" are 2: four (lc3_sns_analyze,
  lc3_sns_spectral_shaping, lc3_tns_analyze, batt_soc_curve_estimate) rest on
  in_stack evidence the corrected rule refutes.

NOT CLOSED
677 entries; 366 image-proven missing-argument sites (untouched -- a WRONG
argument is worse than a missing one); the 2 short definitions; the 303
invented-argument sites; the __ieee754_exp.c u8 collision (still the highest-risk
item); 12 __aeabi_* helpers excluded by rule; 31 refused on 64-bit-return
ambiguity; 17 needing a non-base-C type; 197 frozen files (35 uncompilable, 162
object-drift).

THE FOUR FRAMEBUFFERS WERE NOT RUN AND ARE NOT CLAIMED -- outstanding gate. The
byte-identical flash payload is an INFERENCE that they hold, recorded as such.

recon/refactor/, all net trees and all shared inputs: 0 lines changed.

Detail: recon/analysis/type_disagreement_repair_2.md

+++ recon/app/src/cal_panel_canvas_coord.c
@@ -6,7 +6,7 @@
  *   debug_print      <- FUN_00019c70 @ 0x00019c70
  *   DEBUG_PRINT                         @ 0x0007dda4
  */
-extern uintptr_t get_device_info(void);
+extern int get_device_info(void);
 extern void debug_print(unsigned long, ...);
 extern void DEBUG_PRINT(unsigned long, ...);
