+++ recon/named/k_uptime_get_8.c
@@ -9,7 +9,6 @@
 
 extern uint64_t uptime_ticks_get(void); /* FUN_00086698 @ 0x00086698 */
 
-int64_t k_uptime_get_8(void)
-{
+unsigned long long k_uptime_get_8(void) {
     return (int64_t)((uptime_ticks_get() * UINT64_C(1000)) >> 15);
 }
