+++ recon/readable_sources/app/g1/set_pending_state_flag.c
@@ -9,8 +9,7 @@
  */
 /* Reconstructed FUN_0002eb28 @ 0x2eb28  (parity: 300/300 trials, PROVEN) */
 
-void set_pending_state_flag(unsigned int param_1,int param_2)
-{
+void set_pending_state_flag(unsigned long param_1, unsigned long param_2) {
   if (param_1 < 3) {
     if (param_2 != 0) {
       *(volatile unsigned char *)(((unsigned long)&g_ancs_inbox_flag_current) /*=0x20018da6*/ + param_1) = 1;
