+++ recon/readable_sources/app/g1/st25dv_build_and_write_ndef_records.c
@@ -27,7 +27,7 @@
 extern int st25dv_mailbox_write_with_retry(void *, unsigned int);
 extern unsigned int event_record_init(unsigned char *, unsigned int *);
 extern unsigned int fill_record_type8(unsigned char *, unsigned int *);
-extern unsigned int invoke_optional_op_offset12(unsigned char *, void *);
+extern unsigned long invoke_optional_op_offset12(unsigned char*, void*);
 extern unsigned opt_record_decode(int *, unsigned char *);
 extern int opt_list_encode(int *, int *, int, unsigned);
 extern unsigned int opt_list_append_node(unsigned int *, unsigned char *);
